Update README.md
This is a modern, responsive, and interactive personal portfolio built using React and styled with CSS. The portfolio showcases my skills, projects and contact information in a clean and professional layout. It is designed to give potential employers, clients, and collaborators a clear view of my technical abilities and professional journey.